DC_1/requirements.txt
grpcio>=1.42.0
grpcio-tools>=1.42.0
protobuf>=3.19.0 